\name{06_libreriaR-package}
\alias{06_libreriaR-package}
\alias{06_libreriaR}
\docType{package}
\title{
\packageTitle{06_libreriaR}
}
\description{
\packageDescription{06_libreriaR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{06_libreriaR}
\packageIndices{06_libreriaR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{06_libreriaR}

Maintainer: \packageMaintainer{06_libreriaR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
